+++ app/src/app/views/stix/tactic/tactic-view/tactic-view.component.ts
@@ -22,8 +22,9 @@ export class TacticViewComponent extends StixViewPage implements OnInit {
     ngOnInit() {
         if (this.tactic.firstInitialized) {
             this.tactic.initializeWithDefaultMarkingDefinitions(this.restApiConnector);
+        } else {
+            this.getTechniques();
         }
-        this.getTechniques();
     }
 
     /**